+++ examples/OneBillionRowChallenge.hs
@@ -1,42 +1,38 @@
 {-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE Strict #-}
-{-# LANGUAGE TypeApplications #-}
+{-# LANGUAGE TemplateHaskell #-}
 
 module OneBillionRowChallenge (run) where
 
-import qualified Data.Map as M
 import qualified DataFrame as D
 import qualified DataFrame.Functions as F
 import qualified DataFrame.Lazy as L
 
-import Data.Text (Text)
 import Data.Time
 import DataFrame.Expression.Operators
 import DataFrame.IO.CSV (readSeparated)
-import DataFrame.Schema (Schema (..), schemaType)
+
+$( D.deriveSchemaValuesFromCsvFile
+    "measurement"
+    "../../1brc/measurements.txt"
+ )
 
 run :: IO ()
 run = do
-    let city = F.col @Text "names"
-    let measurement = F.col @Double "temperature"
-    let schema =
-            Schema $
-                M.fromList
-                    [ (F.name city, schemaType @Text)
-                    , (F.name measurement, schemaType @Double)
-                    ]
-
     startCalculation <- getCurrentTime
     df <-
-        L.scanCsvStreamingWith readSeparated schema "../../1brc/measurements.txt"
+        L.scanCsvStreamingWith
+            readSeparated
+            measurementSchema
+            "../../1brc/measurements.txt"
             |> L.groupBy
-                [F.name city]
-                [ F.minimum measurement `as` "minimum"
-                , F.mean measurement `as` "mean"
-                , F.maximum measurement `as` "maximum"
+                [F.name measurementNames]
+                [ F.minimum measurementTemperature `as` "minimum"
+                , F.mean measurementTemperature `as` "mean"
+                , F.maximum measurementTemperature `as` "maximum"
                 ]
             |> L.runDataFrame
-    print $ D.sortBy [D.Asc city] df
+    print $ D.sortBy [D.Asc measurementNames] df
     endCalculation <- getCurrentTime
     let calculationTime = diffUTCTime endCalculation startCalculation
     putStrLn $ "Calculation Time: " ++ show calculationTime